chore: Update tests for DLMM
