+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "outDir": "dist",
     "types": [
-      "jest"
+      "jest",
+      "bun"
     ],
     "typeRoots": [
       "./node_modules/@types"
